+++ app/src/main/kotlin/com/d4rk/androidtutorials/ui/android/textbox/tabs/TextboxTabLayoutFragment.kt
@@ -6,6 +6,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.d4rk.androidtutorials.R
 import com.d4rk.androidtutorials.databinding.FragmentTextboxLayoutBinding
+import me.zhanghai.android.fastscroll.FastScrollerBuilder
 import java.io.ByteArrayOutputStream
 import java.io.IOException
 import java.io.InputStream
@@ -14,6 +15,7 @@ class TextboxTabLayoutFragment : Fragment() {
     private val binding get() = _binding
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
         _binding = FragmentTextboxLayoutBinding.inflate(inflater, container, false)
+        FastScrollerBuilder(binding.textboxScrollView).useMd2Style().build()
         val inputStream: InputStream = resources.openRawResource(R.raw.text_textbox_xml)
         val byteArrayOutputStream = ByteArrayOutputStream()
         var i: Int